Задание 5.14: - подключил файл main.js как точку входа; - создал модули data.js и utils.js
