+++ index.html
@@ -234,7 +234,6 @@
       </section>
     </template>
 
-    <script src="/js/main.js"></script>
-    <script src="/js/functions.js"></script>
+    <script type="module" src="./js/main.js"></script>
 </body>
 </html>
