+++ js/utils.js
@@ -1,0 +1,21 @@
+const getRandomInteger = (a, b) => {
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+};
+
+const getRandomArrayElement = (array) => {
+  const randomIndex = getRandomInteger(0, array.length - 1);
+  return [array[randomIndex], randomIndex];
+};
+
+const generateId = () => {
+  let lastGeneratedId = 0;
+
+  return () => {
+    lastGeneratedId++;
+    return lastGeneratedId;
+  };
+};
+export { getRandomInteger, getRandomArrayElement, generateId };
